Refactor javascript injection proxy
Proxy forks into background and its URL is available
as a label on the container.

A little unfortunate in that we have to perform a health check
for existing containers to verify the proxy is still up
and if not, recreate the entire container as we need
to update the proxy url label. Not sure if there is a better
alternative to this. Maybe just throw an error and have the
user cleanup?